wip: isolated multi_carriage_details safety into module
